requirements.txt
numpy>=1.23,<2
pyannote.audio==2.1.1
pydub==0.25.1
torch==2.8.0
torchaudio==2.8.0
torchvision==0.15.2